changed some floats to doubles

+++ power8tm-pisces/benchmarks/tpcc/code/tpccdb.cc
@@ -17,25 +17,25 @@ void Address::copy(char* street1, char* street2, char* city, char* state, char* 
 
 // Non-integral constants must be defined in a .cc file. Needed for Mac OS X.
 // http://www.research.att.com/~bs/bs_faq2.html#in-class
-const float Item::MIN_PRICE;
-const float Item::MAX_PRICE;
-const float Warehouse::MIN_TAX;
-const float Warehouse::MAX_TAX;
-const float Warehouse::INITIAL_YTD;
-const float District::MIN_TAX;
-const float District::MAX_TAX;
-const float District::INITIAL_YTD;  // different from Warehouse
-const float Customer::MIN_DISCOUNT;
-const float Customer::MAX_DISCOUNT;
-const float Customer::INITIAL_BALANCE;
-const float Customer::INITIAL_CREDIT_LIM;
-const float Customer::INITIAL_YTD_PAYMENT;
+const double Item::MIN_PRICE;
+const double Item::MAX_PRICE;
+const double Warehouse::MIN_TAX;
+const double Warehouse::MAX_TAX;
+const double Warehouse::INITIAL_YTD;
+const double District::MIN_TAX;
+const double District::MAX_TAX;
+const double District::INITIAL_YTD;  // different from Warehouse
+const double Customer::MIN_DISCOUNT;
+const double Customer::MAX_DISCOUNT;
+const double Customer::INITIAL_BALANCE;
+const double Customer::INITIAL_CREDIT_LIM;
+const double Customer::INITIAL_YTD_PAYMENT;
 const char Customer::GOOD_CREDIT[] = "GC";
 const char Customer::BAD_CREDIT[] = "BC";
-const float OrderLine::MIN_AMOUNT;
-const float OrderLine::MAX_AMOUNT;
+const double OrderLine::MIN_AMOUNT;
+const double OrderLine::MAX_AMOUNT;
 const char NewOrderOutput::INVALID_ITEM_STATUS[] = "Item number is not valid";
-const float History::INITIAL_AMOUNT;
+const double History::INITIAL_AMOUNT;
 
 TPCCUndo::~TPCCUndo() {
 }
